+++ noodles-sam/src/header/read_group.rs
@@ -293,13 +293,13 @@ impl ReadGroup {
     ///
     /// let read_group = ReadGroup::builder()
     ///     .set_id("rg0")
-    ///     .insert(Tag::Other(String::from("zn")), String::from("noodles"))
+    ///     .insert(Tag::Other([b'z', b'n']), String::from("noodles"))
     ///     .build()?;
     ///
     /// let fields = read_group.fields();
     /// assert_eq!(fields.len(), 1);
     /// assert_eq!(
-    ///     fields.get(&Tag::Other(String::from("zn"))),
+    ///     fields.get(&Tag::Other([b'z', b'n'])),
     ///     Some(&String::from("noodles"))
     /// );
     ///